pyproject.toml
[project]
name = "live-kit-tuts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[google,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
]
